+++ slashml/utils.py
@@ -10,6 +10,7 @@ def envApiKey():
 
 
 def apiRoot():
+    # return "http://localhost:8000"
     return "https://api.slashml.com"