+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 All notable changes to `slashml-python-client` aka `slashml` will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
+## 0.1.6 - 2023-12-21
+
+### Added
+- Pin some dependencies
+- Add a requirements flag for model deployment
+
+
+## 0.1.5 - 2023-09-17
+
+### Added
+- Unpin depenedencies
+
 ## 0.1.4 - 2023-06-09
 
 ### Added